All cells update

+++ MV-Engine/source/MV/stateSystem/StateSystem.cpp
@@ -31,7 +31,7 @@ namespace mv
 			return false;
 		}
 
-		int8_t number;
+		int number;
 		int colorComponents[3];
 		std::string name;
